% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{epicseg}
\alias{epicseg}
\alias{epicseg-package}
\title{Chromatin segmentation}
\description{
The package provides methods for finding chromatin states in
epigenomic data. There is an R interface as well as a
command line interface.
}
\author{
Alessandro Mammana \email{mammana@molgen.mpg.de}
}

