% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalizecounts.R
\name{normalizecounts}
\alias{normalizecounts}
\title{Normalize a list of count matrices}
\usage{
normalizecounts(clist, normFun = quantileNormalization, nthreads = 1, ...)
}
\arguments{
\item{clist}{A list of count matrices. Each matrix must have the same set of
row names and the same number of columns.}

\item{normFun}{A normalization function to use. Such a function takes a
matrix as input argument and returns a matrix with the same dimensions
as output. The goal of this function is to make the columns comparable.
The functions \code{quantileNormalization} and \code{linearNormalization}
are two already implemented, customizable functions.}

\item{nthreads}{The maximum number of threads. Parallelization is done on the
histone marks using \code{mclapply}. If also \code{normFun} accepts the
\code{nthreads} argument, parallelization will be nested, with the outer
level using \code{min(nthreads, nmarks)} threads, and the inner level
using \code{floor(nthreads/nthreadsOuter)} threads.}

\item{...}{options passed to normFun.}
}
\value{
A list of count matrices similar to the input list, but with the
    rows of each matrix normalized.
}
\description{
Given a list of count matrices, corresponding rows across the different count
matrices are made comparable with a normalization procedure.
Rows correspond if they have the same row name in all count matrices.
}

