% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Produce an html report of a segmentation}
\usage{
report(segments, model, outdir = ".", prefix = "", colors = NULL,
  labels = NULL, annots = list(), rdata = NULL, autoColors = NULL)
}
\arguments{
\item{segments}{GRanges or GRangesList object containing the segmentation. 
If GRanges, the \code{names} slot must be a number from 1 to the maximum 
number of states. Same is true for the unlisted data in a GRangesList.
If a GRangesList, the paths to of each bed file will depend on the names
given by \code{names(segments)}.}

\item{model}{A list with the complete set of the parameters that describe
the HMM, such as that returned by the \code{segment} function.}

\item{outdir}{Output directory where the report will be created.}

\item{prefix}{Prefix prepended to all output files.}

\item{colors}{A character vector assigning one color per state.
each color must be a valid R color.}

\item{labels}{A character vector assigning one name per state}

\item{annots}{Named list where each item is a GRanges object to be
compared with the segmentation.}

\item{rdata}{R objects that will be saved as Rdata archives and will be
part of the report.}

\item{autoColors}{a list with two fields that controls how
the colors are automatically assigned to each state. The value \code{NULL}
disables automatic coloring.}
}
\value{
The path to the newly created webpage
}
\description{
Produce an html report of a segmentation
}
\details{
A web page will be created with plots linked to
    tables in text format.
}

