% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Produce an html report of a segmentation}
\usage{
report(segments, model, counts = NULL, outdir = ".", prefix = "",
  colors = NULL, labels = NULL, annots = list(), rdata = NULL,
  autoColors = defaultAutoColors)
}
\arguments{
\item{segments}{GRanges object containing the segmentation. The \code{names} slot
must be a number from 1 to the maximum number of states.}

\item{model}{A list with the complete set of the parameters that describe
the HMM, such as that returned by the \code{segment} function.}

\item{counts}{Count matrix matching with the 'segments' parameter.
The counts will be used to infer the state colors automatically and to
normalize the mean counts matrix (mean level of a certain mark
in a certain state).}

\item{outdir}{Output directory where the report will be created.}

\item{prefix}{Prefix prepended to all output files.}

\item{colors}{A character vector assigning one color per state.
each color must be a valid R color.}

\item{labels}{A character vector assigning one name per state}

\item{annots}{Named list where each item is a GRanges object to be
compared with the segmentation.}

\item{rdata}{R objects that will be saved as Rdata archives and will be
part of the report.}

\item{autoColors}{a list with two fields that controls how
the colors are automatically assigned to each state.}
}
\value{
The path to the newly created webpage
}
\description{
Produce an html report of a segmentation
}
\details{
A web page will be created with plots linked to
	tables in text format.
}

